solar.js: enhance simulation mode
Enhance the simulation mode for the Smart Solar Panel OCF server.
When running in simulation mode, it does not really simulate any
value; the percentage tilt stays at 0 and the lcd1 and lcd2 values
stay at what they are initially set to.

With this patch, the tilt percentage value goes up and down between
0 and 100% and the time, date and location information are available.
All these are reported by the tiltPercentage, lcd1 and lcd2 properties.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>